requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
blis=0.7.4=pypi_0
ca-certificates=2021.4.13=h06a4308_1
catalogue=1.0.0=pypi_0
certifi=2020.12.5=py36h06a4308_0
chardet=4.0.0=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cymem=2.0.5=pypi_0
einops=0.3.0=pypi_0
en-vectors-web-lg=2.3.0=pypi_0
freetype=2.10.4=h5ab3b9f_0
idna=2.10=pypi_0
importlib-metadata=3.10.1=pypi_0
intel-openmp=2020.2=254
jpeg=9b=h024ee3a_2
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
lz4-c=1.9.3=h2531618_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
murmurhash=1.0.5=pypi_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1k=h27cfd23_0
pillow=8.2.0=py36he98fc37_0
pip=21.0.1=py36h06a4308_0
plac=1.1.3=pypi_0
preshed=3.0.5=pypi_0
python=3.6.13=hdb3f193_0
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pyyaml=5.4.1=py36h27cfd23_1
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
setuptools=52.0.0=py36h06a4308_0
six=1.15.0=py36h06a4308_0
spacy=2.3.5=pypi_0
sqlite=3.35.4=hdfb4753_0
srsly=1.0.5=pypi_0
thinc=7.4.5=pypi_0
tk=8.6.10=hbc83047_0
torchvision=0.7.0=py36_cu102
tqdm=4.60.0=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.4=pypi_0
wasabi=0.8.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
